requirements-tests.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra=tests --output-file=requirements-tests.txt
#
backoff==2.2.1
    # via opentelemetry-exporter-otlp-proto-http
cachetools==5.3.0
    # via
    #   google-auth
    #   wipac-rest-tools
certifi==2022.12.7
    # via
    #   kubernetes
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
coloredlogs==15.0.1
    # via
    #   UNKNOWN (setup.py)
    #   wipac-telemetry
cryptography==40.0.2
    # via pyjwt
dacite==1.8.0
    # via UNKNOWN (setup.py)
deprecated==1.2.13
    # via opentelemetry-api
dnspython==2.3.0
    # via pymongo
exceptiongroup==1.1.1
    # via pytest
google-auth==2.17.3
    # via kubernetes
googleapis-common-protos==1.56.2
    # via
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.53.0
    # via opentelemetry-exporter-jaeger-proto-grpc
humanfriendly==10.0
    # via coloredlogs
idna==3.4
    # via requests
importlib-metadata==6.0.1
    # via opentelemetry-api
iniconfig==2.0.0
    # via pytest
kubernetes==26.1.0
    # via UNKNOWN (setup.py)
motor==3.1.2
    # via UNKNOWN (setup.py)
nest-asyncio==1.5.6
    # via UNKNOWN (setup.py)
oauthlib==3.2.2
    # via requests-oauthlib
opentelemetry-api==1.17.0
    # via
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-jaeger-thrift
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   wipac-telemetry
opentelemetry-exporter-jaeger==1.17.0
    # via wipac-telemetry
opentelemetry-exporter-jaeger-proto-grpc==1.17.0
    # via opentelemetry-exporter-jaeger
opentelemetry-exporter-jaeger-thrift==1.17.0
    # via opentelemetry-exporter-jaeger
opentelemetry-exporter-otlp-proto-http==1.17.0
    # via wipac-telemetry
opentelemetry-proto==1.17.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.17.0
    # via
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-jaeger-thrift
    #   opentelemetry-exporter-otlp-proto-http
    #   wipac-telemetry
opentelemetry-semantic-conventions==0.38b0
    # via opentelemetry-sdk
packaging==23.1
    # via pytest
pluggy==1.0.0
    # via pytest
protobuf==3.20.3
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
    #   wipac-telemetry
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.5.0
    # via wipac-rest-tools
pymongo==4.3.3
    # via
    #   UNKNOWN (setup.py)
    #   motor
pypng==0.20220715.0
    # via qrcode
pytest==7.3.0
    # via
    #   UNKNOWN (setup.py)
    #   pytest-asyncio
    #   pytest-mock
pytest-asyncio==0.21.0
    # via UNKNOWN (setup.py)
pytest-mock==3.10.0
    # via UNKNOWN (setup.py)
python-dateutil==2.8.2
    # via kubernetes
pyyaml==6.0
    # via kubernetes
qrcode==7.4.2
    # via wipac-rest-tools
requests==2.28.2
    # via
    #   UNKNOWN (setup.py)
    #   kubernetes
    #   opentelemetry-exporter-otlp-proto-http
    #   requests-futures
    #   requests-oauthlib
    #   wipac-dev-tools
    #   wipac-rest-tools
requests-futures==1.0.0
    # via wipac-rest-tools
requests-oauthlib==1.3.1
    # via kubernetes
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   kubernetes
    #   python-dateutil
    #   thrift
thrift==0.16.0
    # via opentelemetry-exporter-jaeger-thrift
tomli==2.0.1
    # via pytest
tornado==6.2
    # via
    #   UNKNOWN (setup.py)
    #   wipac-rest-tools
typeguard==3.0.2
    # via UNKNOWN (setup.py)
types-cryptography==3.3.23.2
    # via pyjwt
typing-extensions==4.5.0
    # via
    #   opentelemetry-sdk
    #   qrcode
    #   typeguard
    #   wipac-dev-tools
    #   wipac-telemetry
urllib3==1.26.15
    # via
    #   kubernetes
    #   requests
websocket-client==1.5.1
    # via kubernetes
wipac-dev-tools==1.6.15
    # via
    #   UNKNOWN (setup.py)
    #   wipac-rest-tools
    #   wipac-telemetry
wipac-rest-tools[telemetry]==1.4.18
    # via UNKNOWN (setup.py)
wipac-telemetry==0.2.6
    # via wipac-rest-tools
wrapt==1.15.0
    # via deprecated
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
